Envio alterações finais no Produto

+++ src/sistemadevendas/controller/ProdutoController.java
@@ -2,6 +2,10 @@ package sistemadevendas.controller;
 
 import java.util.List;
 import javax.swing.JOptionPane;
+import sistemadevendas.exceptions.ExistenteException;
+import sistemadevendas.exceptions.FalhaProdutoException;
+import sistemadevendas.exceptions.IdNegativoException;
+import sistemadevendas.exceptions.NaoEncontradoException;
 import sistemadevendas.exceptions.ProdutoExisteException;
 import sistemadevendas.model.Produto;
 import sistemadevendas.services.ProdutoService;
@@ -14,53 +18,48 @@ public class ProdutoController {
     
     private ProdutoService produtoService = new ProdutoService();
     
-    public Produto buscarProdutoPorId(int id){
+    public Produto buscarProdutoPorId(int id) throws IdNegativoException, FalhaProdutoException, NaoEncontradoException{
         
+        if(id<=0){
+            throw new IdNegativoException("O Id não pode ser negativo.");
+        }
         
-        try{
-            Produto p = produtoService.buscarProdutoPorId(id);
-            return p;
+      
+        Produto p = produtoService.buscarProdutoPorId(id);
+        return p;
             
-        } catch(RuntimeException e){
-            JOptionPane.showMessageDialog(null, e.getMessage(), "Erro", JOptionPane.ERROR_MESSAGE);
-            return null;
+
+    }
+    
+    public void adicionarProduto(Produto produto) throws ExistenteException, FalhaProdutoException{
+        if(produto == null){
+            throw new NullPointerException("O produto que está sendo adicionado é nulo");
         }
+        produtoService.adicionarProduto(produto);
         
-        catch(Exception e){
-            return null;
-        }  
     }
     
-    public void adicionarProduto(Produto produto){
-        try{
-            produtoService.adicionarProduto(produto);
-            JOptionPane.showMessageDialog(null, "Produto cadastrado com sucesso no banco de dados");
-        } catch(ProdutoExisteException e) {
-            JOptionPane.showMessageDialog(null, "Não foi possível adicionar: " + e.getMessage());
-        } catch (Exception e){
-            JOptionPane.showMessageDialog(null, "Erro ao adicionar produto: " + e.getMessage());
+    public void removerProduto(int id) throws IdNegativoException, FalhaProdutoException, NaoEncontradoException{
+        if(id<=0){
+            throw new IdNegativoException("O Id não pode ser negativo.");
         }
-    }
-    
-    public void removerProduto(int id){
-        try{
-          produtoService.removerProduto(id);  
-        }catch(Exception e){
-           JOptionPane.showMessageDialog(null, "Houve um erro na hora de remover: " + e.getMessage());
-        }
+        produtoService.removerProduto(id);  
+        
          
     }
     
-    public void editarProduto(Produto produto){
-        try{
-            produtoService.editarProduto(produto);
-        }catch(Exception e){
-            JOptionPane.showMessageDialog(null, "Houve um erro na hora de editar: " + e.getMessage());
+    public void editarProduto(Produto produto) throws IdNegativoException, ExistenteException, FalhaProdutoException, NaoEncontradoException{
+        
+        if(produto.getIdProduto()<=0){
+            throw new IdNegativoException("O Id não pode ser negativo.");
         }
+        
+        produtoService.editarProduto(produto);
+        
  
     }
     
-    public List<Produto> listarProdutos(){
+    public List<Produto> listarProdutos() throws FalhaProdutoException{
         return produtoService.listarProdutos();
     }
 }
